pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gspot_fastapi_auth"
version = "0.0.1"
authors = [
    { name = "gravity48", email = "gravity4849@gmail.com" },
]
description = "GSpot authentication package for FastApi"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'redis >=4.6.0',
    'fastapi >=0.98.0',
    'httpx>=0.24.1',
    'factory-boy>=3.2.1',
    'faker>=18.11.1',
]

[project.urls]
"Homepage" = "https://github.com/gravity48/gspot_fastapi_auth"
"Bug Tracker" = "https://github.com/gravity48/gspot_fastapi_auth/issues"

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''
